requirements.txt
alembic==1.6.0
APScheduler==3.6.3
certifi==2020.12.5
greenlet==1.0.0
importlib-metadata==4.0.1
Mako==1.1.4
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-telegram-bot==13.5
pytz==2021.1
schedule==1.1.0
six==1.15.0
SQLAlchemy==1.4.13
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
zipp==3.4.1
